+++ app/package.json
@@ -17,8 +17,8 @@
     "build-storybook": "build-storybook -s public"
   },
   "dependencies": {
-    "@emotion/core": "10.0.28",
-    "@emotion/styled": "10.0.27",
+    "@emotion/react": "11.4.0",
+    "@emotion/styled": "11.3.0",
     "@improbable-eng/grpc-web": "0.14.0",
     "big.js": "5.2.2",
     "bootstrap": "4.5.0",
@@ -26,7 +26,6 @@
     "d3": "6.3.1",
     "date-fns": "2.14.0",
     "debug": "4.1.1",
-    "emotion-theming": "10.0.27",
     "file-saver": "2.0.2",
     "history": "4.10.1",
     "http-proxy-middleware": "1.0.4",
