+++ app/src/components/pool/orders/OrdersList.tsx
@@ -1,5 +1,6 @@
 import React, { useCallback } from 'react';
 import { observer } from 'mobx-react-lite';
+import { Theme } from '@emotion/react';
 import { usePrefixedTranslation } from 'hooks';
 import { useStore } from 'store';
 import { Order } from 'store/models';
@@ -7,7 +8,7 @@ import { Close, Empty, Scrollable } from 'components/base';
 import SortableHeader from 'components/common/SortableHeader';
 import Tip from 'components/common/Tip';
 import Unit from 'components/common/Unit';
-import { styled, Theme } from 'components/theme';
+import { styled } from 'components/theme';
 import { Table, TableCell, TableHeader, TableRow } from './OrderTable';
 
 /** maps a order status to a theme color */
