+++ app/src/components/tour/TourHost.tsx
@@ -2,9 +2,9 @@
 import React from 'react';
 import Tour, { ReactourStep } from 'reactour';
 import { observer } from 'mobx-react-lite';
-import { useTheme } from 'emotion-theming';
+import { useTheme } from '@emotion/react';
 import { useStore } from 'store';
-import { styled, Theme } from 'components/theme';
+import { styled } from 'components/theme';
 import ChannelListStep from './ChannelListStep';
 import LoopInfoStep from './LoopInfoStep';
 import NodeStatusStep from './NodeStatusStep';
@@ -135,7 +135,7 @@ const Styled = {
 
 const TourHost: React.FC = () => {
   const { appView } = useStore();
-  const theme = useTheme<Theme>();
+  const theme = useTheme();
 
   const { Tour } = Styled;
   return (
